credit-app/requirements.txt.txt
streamlit
pandas
tabula-py